+++ patterns/comments.php
@@ -5,6 +5,7 @@
  * Inserter: no
  *
  * @package Unhurried
+ * @author  Switch Case Studio
  * @since 1.0.0
  */