+++ patterns/posts-grid.php
@@ -7,6 +7,7 @@
  * Viewport width: 1400
  *
  * @package Unhurried
+ * @author  Switch Case Studio
  * @since 1.0.0
  */